requirements.txt
discord.py[voice]
python-dotenv
google-api-python-client
gTTS
yt-dlp
openai
PyNaCl